refactor(gui): thin arena scene panel
